Fix for Issue 873 : Replication issue with handlerServlet (http://code.google.com/p/mobicents/issues/detail?id=873)
When the subsequent routing is trying to find the corresponding sip session to the message.
It does a first try and if it doesn't find anything it tries in inverting the from and to (since the message can come from the other side), the HA code was creating and replicating a new sip session even if the first lookup didn't find anything leading to the NPE on handlerServlet later on when another subsequent message is received.
We don't create any new sip session now which should result in a little perf improvement as well for those cases

git-svn-id: https://mobicents.googlecode.com/svn/trunk/servers/sip-servlets@7715 bf0df8d0-2c1f-0410-b170-bd30377b63dc