Issue 965 : Move Sip Session Dialog replication to JAIN SIP HA (http://code.google.com/p/mobicents/issues/detail?id=965)
lazy loading of the sip session in the Sip Servlet message and transient now so that it can be replicated.
lazy loading of the sip app session in the sip session impl
making the sipfactory replicable
removing the replication of the dialog when replication the sip session, just replicating the dialogId.
Fixes for Testsuite regressions

git-svn-id: https://mobicents.googlecode.com/svn/trunk/servers/sip-servlets@7786 bf0df8d0-2c1f-0410-b170-bd30377b63dc