Update Issue 2130
Committing a fix to avoid memory leak of the Client transaction when there is an error sending the request out.

We take 2 approaches, throwing an IllegalStateException before creating the client transaction if the dialog is in TERMINATED state (thus avoid consuming resources for nothing) and in case a race condition occurs between sending the BYE and the INFO, we terminate the client transaction if the request couldn't be sent out so that it doesn't leak.

git-svn-id: https://mobicents.googlecode.com/svn/trunk/servers/sip-servlets@15637 bf0df8d0-2c1f-0410-b170-bd30377b63dc