Update Issue 1872
gravitate the timers where the LB decided to fail over the subsequent request for a sipapplication session which would work better with any kind of Load Balancer but it is done only if the concurrency control mode is SipApplicationSession.basically the timer is cancelled remotely and rescheduled locally. the most recent node that recreate a sip application session will move the timers locally in recreating the sip application session  
if the sip application session is present on multiple nodes because there is no affinity then the timers will sit where the sipappsession was recreated last (note recreated, not updated).

it would stay at the node where it was recreated last it the sip application session flip flop due to LB flip flopping to different nodes if there is no affinity then there is no guarantees anyway wrt to concurrency control
so no need to flip flop the timers

git-svn-id: https://mobicents.googlecode.com/svn/trunk/servers/sip-servlets@15685 bf0df8d0-2c1f-0410-b170-bd30377b63dc