Update Issue 2264
Georges I found some issues with the testcases so I updated some things and added an updateTimer on ProxyBranchImpl on 200 OK to PRACK and UPDATE since there is activity on the branch, timer C has to be reset.
Otherwise, looks good, so committing.
Horaci can you verify it ?

Thanks
Jean

git-svn-id: https://mobicents.googlecode.com/svn/trunk/servers/sip-servlets@17318 bf0df8d0-2c1f-0410-b170-bd30377b63dc