Update Issue 2450
Reactoring the concurrency control to deal with this case.
We are now keeping a set of application sessions created/accessed in the application code in a thread local and we make sure to release all sip application sessions when the application code is exited so that there is no concurrent access possible on a sip application session that has been created in the application code and is accessed at the same time by another container managed thread.

git-svn-id: https://mobicents.googlecode.com/svn/trunk/servers/sip-servlets@17727 bf0df8d0-2c1f-0410-b170-bd30377b63dc