Update Issue 2474
Fix to impl and spec.
Non regression test showcasing testing JSR 289 Section 10.2.4.2 Correlating responses to proxy branches 
The handling is now as follow : "doBranchResponse() is called for intermediate branch final responses, and doSuccessResponse() or doErrorResponse() is called for just the best final response. If doBranchResponse() is not overridden, only the best result is seen. A difference would only be seen if doResponse() itself was overridden."

git-svn-id: https://mobicents.googlecode.com/svn/trunk/servers/sip-servlets@17753 bf0df8d0-2c1f-0410-b170-bd30377b63dc