Update Issue 2354
Fix + non regression test.

We allow to forward provisional responses (to play mandatory early media by example) and create derived session for it. But forwarding 200 OK is not allowed as per RFC3261 and TCK test com.bea.sipservlet.tck.agents.api.javax_servlet_sip.B2buaHelperTest.testCreateResponseToOriginalRequest101.

The decision has to be made at the B2BUAHelper when a 200 OK has already been forwarded and the B2BUAHelper should CANCEL other branches upon receiving a 200 OK. If there is a race condition and another endpoint sent a 200 OK before we had the time to cancel the leg then it should be ACK-ed and BYE-ed


git-svn-id: https://mobicents.googlecode.com/svn/trunk/servers/sip-servlets@17838 bf0df8d0-2c1f-0410-b170-bd30377b63dc